refactor(registry): split into submodules
src/registry.rs grew to 1173 lines. Split into a module directory
with focused files: types (struct + builder), checks (per-linter
constructors), obsolete (legacy key migration), mise (tool discovery
and activation), resolve (binary lookup), tests (unchanged). Public
API preserved via re-exports in mod.rs.
