+++ src/registry/mod.rs
@@ -1,0 +1,14 @@
+mod checks;
+mod mise;
+mod obsolete;
+mod resolve;
+mod types;
+
+pub use checks::builtin;
+pub use mise::{check_active, read_mise_tools};
+pub use obsolete::{OBSOLETE_KEYS, find_obsolete_key};
+pub use resolve::{binary_on_path, resolve_bin_name};
+pub use types::{Category, Check, CheckKind, Scope, SpecialKind};
+
+#[cfg(test)]
+mod tests;
