+++ src/registry/types.rs
@@ -1,0 +1,332 @@
+/// How a check is invoked relative to the file list.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum Scope {
+    /// Invoked once per matched file: `{FILE}` placeholder.
+    File,
+    /// Invoked once with all matched files: `{FILES}` placeholder.
+    Files,
+    /// Invoked once with no file args (e.g. golangci-lint).
+    Project,
+}
+
+/// Which init profile (and `--fast-only` behaviour) a check belongs to.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Default)]
+pub enum Category {
+    /// Primary programming language linter/formatter (Rust, Python, Go, …) — all init profiles.
+    Lang,
+    /// Supplementary language check (shell, Docker, CI/CD) — `default` + `comprehensive` only.
+    Style,
+    /// General fast tool (not language-specific) — `default` and `comprehensive` init profiles.
+    #[default]
+    Default,
+    /// Slow tool — `comprehensive` init profile only; skipped when `--fast-only` is passed.
+    Slow,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum SpecialKind {
+    Links,
+    RenovateDeps,
+    LicenseHeader,
+}
+
+#[derive(Debug, Clone)]
+pub enum CheckKind {
+    Template {
+        check_cmd: &'static str,
+        fix_cmd: &'static str,
+        /// When set and `file_list.full == true`, used instead of `check_cmd` as a
+        /// project-wide check command (no `{FILES}` substitution). Useful for tools like
+        /// `cargo fmt` that handle all-files scanning better than a long file list.
+        full_cmd: &'static str,
+        /// When set and `file_list.full == true` in fix mode, used instead of `fix_cmd`.
+        full_fix_cmd: &'static str,
+        scope: Scope,
+    },
+    Special(SpecialKind),
+}
+
+#[derive(Debug, Clone)]
+pub struct Check {
+    pub name: &'static str,
+    /// Binary name used to invoke the tool.
+    pub bin_name: &'static str,
+    /// mise.toml tool key to look up for availability. When `None`, falls back to
+    /// `bin_name`. Use this when the binary comes from a toolchain entry rather than
+    /// its own tool entry (e.g. `cargo-fmt` ships with `rust`).
+    pub mise_tool_name: Option<&'static str>,
+    /// Semver requirement string (e.g. `">=1.0.0"`). When `None`, any version matches.
+    /// When multiple registry entries share a `bin_name`, each must have a `version_range`
+    /// and the ranges must be non-overlapping and collectively exhaustive.
+    pub version_range: Option<&'static str>,
+    /// Glob patterns for matching files.
+    pub patterns: &'static [&'static str],
+    /// When any of these named checks are active, exclude their patterns from
+    /// this check's file list. Used to avoid double-checking files that a
+    /// dedicated formatter already owns.
+    pub excludes_if_active: &'static [&'static str],
+    pub category: Category,
+    /// When set, look for `(filename, flag)` in config_dir: if the file exists, inject
+    /// `flag <abs-path>` into the command right after the binary name.
+    pub linter_config: Option<(&'static str, &'static str)>,
+    /// This check is a formatter — it owns certain file types for formatting purposes.
+    pub is_formatter: bool,
+    /// Skip files owned by active formatters (used by ec to avoid double-checking).
+    pub defers_to_formatters: bool,
+    /// Always considered active regardless of mise.toml (used for config-activated checks).
+    pub activate_unconditionally: bool,
+    /// Canonical mise tool key to write when setting up a new project (e.g. `npm:prettier`).
+    /// Optional mise toolchain components to request when installing via `flint init`
+    /// (e.g. `"clippy,rustfmt"` for the `rust` toolchain). Produces an inline-table
+    /// entry: `rust = { version = "latest", components = "clippy,rustfmt" }`.
+    pub mise_install_components: Option<&'static str>,
+    /// On Windows, the binary is a self-executing JAR that cannot be run directly
+    /// or via cmd.exe — invoke as `java -jar <resolved-path>` instead.
+    pub windows_java_jar: bool,
+    pub kind: CheckKind,
+    /// Binary name format when the backend installs with a versioned name (e.g. `"shfmt_{version}"`
+    /// → `"shfmt_v3.12.0"`). `{version}` is replaced with the version declared in mise.toml.
+    /// Paired with `mise_tool_name` when the backend names binaries with a version suffix.
+    pub versioned_bin_fmt: Option<&'static str>,
+    /// Plain-text description of what the check does — shown in `flint linters` and the README table.
+    pub desc: &'static str,
+    /// Extended markdown documentation shown in the README detail section (behaviour, config examples).
+    pub docs: &'static str,
+}
+
+impl Check {
+    pub fn has_fix(&self) -> bool {
+        match &self.kind {
+            CheckKind::Template { fix_cmd, .. } => !fix_cmd.is_empty(),
+            CheckKind::Special(SpecialKind::Links) => false,
+            CheckKind::Special(SpecialKind::RenovateDeps) => true,
+            CheckKind::Special(SpecialKind::LicenseHeader) => false,
+        }
+    }
+
+    /// Returns false for checks implemented entirely in-process with no external binary.
+    pub fn uses_binary(&self) -> bool {
+        !matches!(self.kind, CheckKind::Special(SpecialKind::LicenseHeader))
+    }
+
+    // --- Constructors ---
+
+    /// Check invoked once per matched file (`{FILE}`). `name` is also used as `bin_name`.
+    pub fn file(
+        name: &'static str,
+        check_cmd: &'static str,
+        patterns: &'static [&'static str],
+    ) -> Self {
+        Self::template(name, patterns, check_cmd, Scope::File)
+    }
+
+    /// Check invoked once with all matched files (`{FILES}`). `name` is also used as `bin_name`.
+    pub fn files(
+        name: &'static str,
+        check_cmd: &'static str,
+        patterns: &'static [&'static str],
+    ) -> Self {
+        Self::template(name, patterns, check_cmd, Scope::Files)
+    }
+
+    /// Check invoked once per project (no file args). `name` is also used as `bin_name`.
+    pub fn project(
+        name: &'static str,
+        check_cmd: &'static str,
+        patterns: &'static [&'static str],
+    ) -> Self {
+        Self::template(name, patterns, check_cmd, Scope::Project)
+    }
+
+    fn template(
+        name: &'static str,
+        patterns: &'static [&'static str],
+        check_cmd: &'static str,
+        scope: Scope,
+    ) -> Self {
+        Check {
+            name,
+            bin_name: name,
+            mise_tool_name: None,
+            version_range: None,
+            patterns,
+            excludes_if_active: &[],
+            linter_config: None,
+            is_formatter: false,
+            defers_to_formatters: false,
+            activate_unconditionally: false,
+            category: Category::Default,
+            mise_install_components: None,
+            kind: CheckKind::Template {
+                check_cmd,
+                fix_cmd: "",
+                full_cmd: "",
+                full_fix_cmd: "",
+                scope,
+            },
+            windows_java_jar: false,
+            versioned_bin_fmt: None,
+            desc: "",
+            docs: "",
+        }
+    }
+
+    /// Special check with custom logic (not a simple command template).
+    pub fn special(name: &'static str, bin_name: &'static str, kind: SpecialKind) -> Self {
+        Check {
+            name,
+            bin_name,
+            mise_tool_name: None,
+            version_range: None,
+            patterns: &[],
+            excludes_if_active: &[],
+            linter_config: None,
+            is_formatter: false,
+            defers_to_formatters: false,
+            activate_unconditionally: false,
+            category: Category::Default,
+            mise_install_components: None,
+            windows_java_jar: false,
+            kind: CheckKind::Special(kind),
+            versioned_bin_fmt: None,
+            desc: "",
+            docs: "",
+        }
+    }
+
+    // --- Modifiers ---
+
+    /// Override `bin_name` when the binary name differs from the check name
+    /// (e.g. `ruff-format` invokes `ruff`).
+    pub fn bin(mut self, bin_name: &'static str) -> Self {
+        self.bin_name = bin_name;
+        self
+    }
+
+    /// Set the mise.toml tool key when the binary ships as part of a toolchain
+    /// (e.g. `cargo-fmt` ships with `rust`).
+    pub fn mise_tool(mut self, name: &'static str) -> Self {
+        self.mise_tool_name = Some(name);
+        self
+    }
+
+    /// Add a fix command (auto-fix mode).
+    pub fn fix(mut self, fix_cmd: &'static str) -> Self {
+        if let CheckKind::Template {
+            fix_cmd: ref mut f, ..
+        } = self.kind
+        {
+            *f = fix_cmd;
+        }
+        self
+    }
+
+    /// Set project-wide commands used instead of `check_cmd`/`fix_cmd` when
+    /// `file_list.full == true` (explicit `--full` or no merge base). Commands
+    /// run with no file arguments — useful for tools that discover files internally
+    /// (e.g. `cargo fmt`). Also handles edition detection for Rust tools.
+    pub fn full_cmd(mut self, check: &'static str, fix: &'static str) -> Self {
+        if let CheckKind::Template {
+            full_cmd: ref mut c,
+            full_fix_cmd: ref mut f,
+            ..
+        } = self.kind
+        {
+            *c = check;
+            *f = fix;
+        }
+        self
+    }
+
+    /// Restrict activation to a semver range of the declared tool version.
+    #[allow(dead_code)]
+    pub fn version_req(mut self, range: &'static str) -> Self {
+        self.version_range = Some(range);
+        self
+    }
+
+    /// On Windows, invoke this binary via `java -jar <path>` rather than directly.
+    /// Use for self-executing JARs (e.g. ktlint) that cmd.exe cannot run.
+    pub fn windows_java_jar(mut self) -> Self {
+        self.windows_java_jar = true;
+        self
+    }
+
+    /// Mark as slow — skipped when `--fast-only` is passed; `comprehensive` init profile only.
+    pub fn slow(mut self) -> Self {
+        self.category = Category::Slow;
+        self
+    }
+
+    /// Mark as a formatter — files it owns are excluded from ec when both are active.
+    pub fn formatter(mut self) -> Self {
+        self.is_formatter = true;
+        self
+    }
+
+    /// Skip files owned by active formatters (for ec — avoids double-checking).
+    pub fn defer_to_formatters(mut self) -> Self {
+        self.defers_to_formatters = true;
+        self
+    }
+
+    /// Always considered active regardless of mise.toml (for config-activated checks).
+    pub fn activate_unconditionally(mut self) -> Self {
+        self.activate_unconditionally = true;
+        self
+    }
+
+    /// Set a versioned binary name format for tools where the backend installs with a
+    /// version suffix (e.g. `"shfmt_{version}"` → `"shfmt_v3.12.0"`). Paired with
+    /// `.mise_tool()` to identify which key provides the version.
+    pub fn versioned_bin(mut self, fmt: &'static str) -> Self {
+        self.versioned_bin_fmt = Some(fmt);
+        self
+    }
+
+    /// Set the plain-text description shown in `flint linters` and the README table.
+    pub fn desc(mut self, desc: &'static str) -> Self {
+        self.desc = desc;
+        self
+    }
+
+    /// Set extended markdown documentation shown in the README detail section.
+    pub fn docs(mut self, docs: &'static str) -> Self {
+        self.docs = docs;
+        self
+    }
+
+    /// Override the patterns field (useful for Special checks that need init-detection
+    /// patterns but don't use them for file matching at runtime).
+    pub fn patterns(mut self, patterns: &'static [&'static str]) -> Self {
+        self.patterns = patterns;
+        self
+    }
+
+    /// Mark as a primary language analysis check — included in all init profiles.
+    pub fn lang(mut self) -> Self {
+        self.category = Category::Lang;
+        self
+    }
+
+    /// Mark as a language-specific style/formatter check — included in all init profiles.
+    pub fn style(mut self) -> Self {
+        self.category = Category::Style;
+        self
+    }
+
+    /// Set toolchain components required when installing via `flint init`
+    /// (e.g. `"clippy,rustfmt"` for the `rust` toolchain).
+    pub fn install_components(mut self, components: &'static str) -> Self {
+        self.mise_install_components = Some(components);
+        self
+    }
+
+    /// Inject a config file from config_dir into the linter command.
+    /// If `config_dir/file` exists at runtime, `flag <abs-path>` is inserted
+    /// right after the binary name. Has no effect when the file is absent.
+    pub fn linter_config(mut self, file: &'static str, flag: &'static str) -> Self {
+        self.linter_config = Some((file, flag));
+        self
+    }
+}
